requirements.txt
fastapi==0.100.0         
uvicorn==0.23.1          
neo4j==5.11.0            
pydantic==2.9.2          
pydantic-settings==2.6.1 
python-dotenv==1.0.0     
